+++ tests/HttpStatusTest.php
@@ -5,81 +5,12 @@ namespace Narrowspark\HttpStatus\Tests;
 use DOMDocument;
 use DomXPath;
 use Narrowspark\HttpStatus\Exception;
+use Narrowspark\HttpStatus\Contract\Exception\HttpException as HttpExceptionContract;
 use Narrowspark\HttpStatus\HttpStatus;
 use PHPUnit\Framework\TestCase;
-use Throwable;
 
 class HttpStatusTest extends TestCase
 {
-    private $errorNames = [
-        // Informational 1xx
-        100 => 'Continue',
-        101 => 'Switching Protocols',
-        102 => 'Processing',
-        // Successful 2xx
-        200 => 'OK',
-        201 => 'Created',
-        202 => 'Accepted',
-        203 => 'Non-Authoritative Information',
-        204 => 'No Content',
-        205 => 'Reset Content',
-        206 => 'Partial Content',
-        207 => 'Multi-Status',
-        208 => 'Already Reported',
-        226 => 'IM Used',
-        // Redirection 3xx
-        300 => 'Multiple Choices',
-        301 => 'Moved Permanently',
-        302 => 'Found',
-        303 => 'See Other',
-        304 => 'Not Modified',
-        305 => 'Use Proxy',
-        306 => '(Unused)',
-        307 => 'Temporary Redirect',
-        308 => 'Permanent Redirect',
-        // Client Error 4xx
-        400 => 'Bad Request',
-        401 => 'Unauthorized',
-        402 => 'Payment Required',
-        403 => 'Forbidden',
-        404 => 'Not Found',
-        405 => 'Method Not Allowed',
-        406 => 'Not Acceptable',
-        407 => 'Proxy Authentication Required',
-        408 => 'Request Timeout',
-        409 => 'Conflict',
-        410 => 'Gone',
-        411 => 'Length Required',
-        412 => 'Precondition Failed',
-        413 => 'Payload Too Large',
-        414 => 'URI Too Long',
-        415 => 'Unsupported Media Type',
-        416 => 'Range Not Satisfiable',
-        417 => 'Expectation Failed',
-        418 => 'I\'m a teapot',
-        421 => 'Misdirected Request',
-        422 => 'Unprocessable Entity',
-        423 => 'Locked',
-        424 => 'Failed Dependency',
-        426 => 'Upgrade Required',
-        428 => 'Precondition Required',
-        429 => 'Too Many Requests',
-        431 => 'Request Header Fields Too Large',
-        451 => 'Unavailable For Legal Reasons',
-        // Server Error 5xx
-        500 => 'Internal Server Error',
-        501 => 'Not Implemented',
-        502 => 'Bad Gateway',
-        503 => 'Service Unavailable',
-        504 => 'Gateway Timeout',
-        505 => 'HTTP Version Not Supported',
-        506 => 'Variant Also Negotiates',
-        507 => 'Insufficient Storage',
-        508 => 'Loop Detected',
-        510 => 'Not Extended',
-        511 => 'Network Authentication Required',
-    ];
-
     private $errorPhrases = [
         // Successful 2xx
         200 => 'Standard response for successful HTTP requests.',
@@ -188,31 +119,20 @@ class HttpStatusTest extends TestCase
         511 => Exception\NetworkAuthenticationRequiredException::class,
     ];
 
-    public function testGetReasonPhrase()
-    {
-        foreach ($this->errorNames as $code => $text) {
-            self::assertSame(
-                $text,
-                HttpStatus::getReasonPhrase($code),
-                'Expected HttpStatus::getReasonPhrase(' . $code . ') to return ' . $text
-            );
-        }
-    }
-
     public function testGetReasonMessage()
     {
         foreach ($this->errorPhrases as $code => $text) {
             self::assertSame(
                 $text,
                 HttpStatus::getReasonMessage($code),
-                'Expected HttpStatus::getReasonPhrase(' . $code . ') to return ' . $text
+                'Expected HttpStatus::getReasonMessage(' . $code . ') to return ' . $text
             );
         }
     }
 
     /**
-     * @expectedException \InvalidArgumentException
-     * @expectedExceptionMessage The submitted code must be a positive integer between 100 and 599.
+     * @expectedException \Narrowspark\HttpStatus\Exception\InvalidArgumentException
+     * @expectedExceptionMessage The submitted code "700" must be a positive integer between 100 and 599.
      */
     public function testGetReasonPhraseToThrowInvalidArgumentException()
     {
@@ -220,7 +140,7 @@ class HttpStatusTest extends TestCase
     }
 
     /**
-     * @expectedException \OutOfBoundsException
+     * @expectedException \Narrowspark\HttpStatus\Exception\OutOfBoundsException
      * @expectedExceptionMessage Unknown http status code: `509`.
      */
     public function testGetReasonPhraseToThrowOutOfBoundsException()
@@ -229,8 +149,8 @@ class HttpStatusTest extends TestCase
     }
 
     /**
-     * @expectedException \InvalidArgumentException
-     * @expectedExceptionMessage The submitted code must be a positive integer between 100 and 599.
+     * @expectedException \Narrowspark\HttpStatus\Exception\InvalidArgumentException
+     * @expectedExceptionMessage The submitted code "700" must be a positive integer between 100 and 599.
      */
     public function testGetReasonMessageToThrowInvalidArgumentException()
     {
@@ -238,7 +158,7 @@ class HttpStatusTest extends TestCase
     }
 
     /**
-     * @expectedException \OutOfBoundsException
+     * @expectedException \Narrowspark\HttpStatus\Exception\OutOfBoundsException
      * @expectedExceptionMessage Unknown http status code: `509`.
      */
     public function testGetReasonMessageToThrowOutOfBoundsException()
@@ -246,43 +166,26 @@ class HttpStatusTest extends TestCase
         HttpStatus::getReasonMessage(509);
     }
 
-    public function testGetReasonException()
-    {
-        foreach ($this->errorNames as $code => $text) {
-            try {
-                HttpStatus::getReasonException($code);
-            } catch (Throwable $exception) {
-                self::assertSame(
-                    $code . ' ' . $text,
-                    $exception->getMessage(),
-                    'Expected HttpStatus::getReasonException(' . $code . ') to return ' . $text
-                );
-
-                self::assertSame(
-                    $code,
-                    $exception->getCode(),
-                    'Expected HttpStatus::getReasonException(' . $code . ') to return ' . $text
-                );
-            }
-        }
-    }
-
-    public function testIfAllExceptionsAreExtendetFromTheRightClient()
+    public function testIfAllExceptionsAreExtendedFromTheRightClient()
     {
         $clientCount = 0;
         $serverCount = 0;
 
-        foreach ($this->errorNames as $code => $text) {
+        foreach ($this->ianaCodesReasonPhrasesProvider() as $data) {
             try {
-                HttpStatus::getReasonException($code);
+                HttpStatus::getReasonException((int) $data[0]);
             } catch (Exception\AbstractClientErrorException $client) {
+                self::assertInstanceOf($this->phrasesExceptions[$data[0]], $client);
+
                 ++$clientCount;
             } catch (Exception\AbstractServerErrorException $server) {
+                self::assertInstanceOf($this->phrasesExceptions[$data[0]], $server);
+
                 ++$serverCount;
             }
         }
 
-        self::assertSame(28, $clientCount);
+        self::assertSame(27, $clientCount);
         self::assertSame(11, $serverCount);
     }
 
@@ -294,7 +197,11 @@ class HttpStatusTest extends TestCase
      */
     public function testReasonPhraseDefaultsAgainstIana($code, $reasonPhrase)
     {
-        self::assertEquals($reasonPhrase, HttpStatus::getReasonPhrase((int) $code));
+        self::assertEquals(
+            $reasonPhrase,
+            HttpStatus::getReasonPhrase((int) $code),
+            'Expected HttpStatus::getReasonPhrase(' . $code . ') to return ' . $reasonPhrase
+        );
     }
 
     /**
@@ -312,23 +219,14 @@ class HttpStatusTest extends TestCase
 
         try {
             HttpStatus::getReasonException((int) $code);
-        } catch (Throwable $exception) {
-            self::assertSame(
-                $code . ' ' . $reasonPhrase,
-                $exception->getMessage(),
-                'Expected HttpStatus::getReasonException(' . $code . ') to return ' . $reasonPhrase
-            );
-
-            self::assertSame(
-                (int) $code,
-                $exception->getCode(),
-                'Expected HttpStatus::getReasonException(' . $code . ') to return ' . $reasonPhrase
-            );
+        } catch (HttpExceptionContract $exception) {
+            self::assertSame($code . ' ' . $reasonPhrase, $exception->getMessage());
+            self::assertSame((int) $code, $exception->getStatusCode());
         }
     }
 
     /**
-     * @link      http://github.com/zendframework/zend-diactoros for the canonical source repository
+     * @link http://github.com/zendframework/zend-diactoros for the canonical source repository
      *
      * @author    Fábio Pacheco
      * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)