+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
 All Notable changes to `http-status` will be documented in this file
+## NEXT - 2017-08-09
+
+### Added
+- fixes #23 Add header functions to exceptions
+- new HttpException Interface for better exception catching
+- added new getStatusCode function to get the http status code
+
+### Deprecated
+- Nothing
+
+### Fixed
+- getCode returns now the exception code and not the http status code
+
+### Removed
+- duplicated tests
+
+### Security
+- Nothing
+
 ## NEXT - 2017-02-20
 
 ### Added